feat: infrastructure overhaul & CI/CD pipeline setup
- Upgraded database to PostgreSQL 18.3 and fixed volume compatibility issues.
- Optimized Dockerfile with multi-stage builds and automated testing.
- Resolved 'pg_ctlcluster' directory errors by implementing a clean volume strategy.
- Implemented GitHub Actions workflow for automated Build, Test, and Push.
- Integrated Docker Hub for centralized image management.
- Configured production-ready environment settings for Constructix API.

+++ src/Core/Constructix.Domain/Common/Guard.cs
@@ -23,7 +23,7 @@ public static class Guard
     }
 
     // 4. Sıfır və ya mənfi yoxlaması (int/double/decimal üçün istifadə oluna bilər)
-    public static void AgainstZeroOrNegative(double value, string name)
+    public static void AgainstZeroOrNegative(decimal value, string name)
     {
         if (value <= 0)
             throw new ArgumentOutOfRangeException(name, $"{name} sıfır və ya mənfi ola bilməz.");
